backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
pinecone
mistralai
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pandas
google-auth-oauthlib
google-api-python-client
email-validator




